requirements.txt
streamlit==1.34.0
streamlit-webrtc==0.46.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.9.0.80
numpy==1.25.2
Pillow==10.2.0
av==10.0.0
